Correctly convert Heroku API responses to JSON
Heroku sends back a Content-Type header indicating that it's responding with
JSON. Its client reads that header and auto-converts the response body from a
string of JSON to a Ruby representation.

Excon does not auto-convert to a Ruby hash, so when we catch an error and grab
`error.response.body`, it's the raw JSON string.

After this change, the HerokuClient always has a `json_body` that is a Ruby
representation of the JSON.